+++ src/finaletoolkit/utils/__init__.py
@@ -1,3 +1,3 @@
 from finaletoolkit.utils.utils import *
-from finaletoolkit.utils._filter_bam import filter_bam
+from finaletoolkit.utils._filter_file import filter_file
 from finaletoolkit.utils._agg_bw import agg_bw